backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
sqlalchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.6.4
pydantic-settings==2.2.1
redis==5.0.3
python-dotenv==1.0.1
litellm==1.34.11
openai>=2.32.0
python-multipart==0.0.9
boto3==1.34.79
pytest==8.1.1
ruff==0.3.5
black==24.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
email-validator>=2.0.0
pypdf==4.2.0
python-docx==1.1.2
